+++ lib/couch_potato/database.rb
@@ -32,6 +32,9 @@ module CouchPotato
         spec.reduce_function).query_view!(spec.view_parameters)
       processed_results = spec.process_results results
       processed_results.instance_eval "def total_rows; #{results['total_rows']}; end" if results['total_rows']
+      processed_results.each do |document|
+        document.database = self if document.respond_to?(:database=)
+      end if processed_results.respond_to?(:each)
       processed_results
     end