+++ spec/unit/database_spec.rb
@@ -205,3 +205,32 @@ describe CouchPotato::Database, 'save_document' do
     
   end
 end
+
+describe CouchPotato::Database, 'view' do
+  before(:each) do
+    @db = CouchPotato::Database.new(stub('couchrest db').as_null_object)
+    @result = stub('result')
+    @spec = stub('view spec', :process_results => [@result]).as_null_object
+    CouchPotato::View::ViewQuery.stub(:new => stub('view query', :query_view! => {'rows' => [@result]}))
+  end
+  
+  it "should set itself on returned results that have an accessor" do
+    @result.stub(:respond_to?).with(:database=).and_return(true)
+    @result.should_receive(:database=).with(@db)
+    @db.view(@spec)
+  end
+  
+  it "should not set itself on returned results that don't have an accessor" do
+    @result.stub(:respond_to?).with(:database=).and_return(false)
+    @result.should_not_receive(:database=).with(@db)
+    @db.view(@spec)
+  end
+  
+  it "should not try to set itself on result sets that are not collections" do
+    lambda {
+      @spec.stub(:process_results => 1)
+    }.should_not raise_error
+    
+    @db.view(@spec)
+  end
+end